--- scalameter-core/js/src/main/scala/org/scalameter/ScalaRuntime.scala
@@ -1,6 +1,0 @@
-package org.scalameter
-
-object ScalaRuntime extends AbstractScalaRuntime {
-
-  def machineContext = Context.empty
-}
