--- scalameter-core/js/src/main/scala/org/scalameter/utils/SystemUtils.scala
@@ -1,9 +1,0 @@
-package org.scalameter
-package utils
-
-/**
- *  SystemUtils for scalaJs
- */
-object SystemUtils extends AbstractSystemUtils {
-  def IS_OS_WINDOWS : Boolean = false
-}
