--- scalameter-core/shared/src/main/scala/org/scalameter/AbstractScalaRuntime.scala
@@ -1,5 +1,0 @@
-package org.scalameter
-
-trait AbstractScalaRuntime {
-  def machineContext : Context 
-}
