--- scalameter-core/shared/src/main/scala/org/scalameter/utils/AbstractSystemUtils.scala
@@ -1,9 +1,0 @@
-package org.scalameter
-package utils
-
-/**
- * Abstraction of the SystemUtils library to detect the OS
- */
-trait AbstractSystemUtils {
-  def IS_OS_WINDOWS : Boolean
-}
